+++ serial.c
@@ -5,18 +5,18 @@
 #include "defs.h"
 
 
-void eval_serial(int* image) {
+void eval_serial(uint8_t* image) {
   Hist n = {};
 
   fasttime_t start = gettime();
   // Collect
   for (int i = 0; i < SER_ITER; i++) {
     for (int j = 0; j < PAR_ITER; j++) {
-      int pixel = image[j];
+      uint8_t pixel = image[j];
       n.ele[pixel]++;
     }
   }
   fasttime_t stop = gettime();
 
-  printf("%f\t%d\t%d\t%d\n", tdiff_sec(start, stop), n.ele[0], n.ele[1], n.ele[255]);
+  printf("%f\t%u\t%u\t%u\n", tdiff_sec(start, stop), n.ele[0], n.ele[1], n.ele[255]);
 }
